Combine the class to make it easier to know the relationship between classes.
