+++ osu.Game.Rulesets.Karaoke/Screens/Edit/Beatmaps/Lyrics/Content/Compose/DetailLyricList.cs
@@ -4,6 +4,9 @@
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Shapes;
+using osu.Game.Rulesets.Karaoke.Objects;
+using osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Lyrics.CaretPosition;
+using osuTK;
 
 namespace osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Lyrics.Content.Compose;
 
@@ -25,7 +28,7 @@ public partial class DetailLyricList : LyricList
         bool containsHandler = mode == LyricEditorMode.EditText;
 
         const float timing_base_width = LYRIC_LIST_PADDING + DetailRow.TIMING_WIDTH;
-        float timingWidth = containsHandler ? DrawableLyricListItem.HANDLER_WIDTH + timing_base_width : timing_base_width;
+        float timingWidth = containsHandler ? HANDLER_WIDTH + timing_base_width : timing_base_width;
         return new GridContainer
         {
             ColumnDimensions = new[]
@@ -50,5 +53,39 @@ public partial class DetailLyricList : LyricList
                 },
             },
         };
+    }
+
+    public partial class DrawableDetailLyricList : DrawableLyricList
+    {
+        protected override Vector2 Spacing => new();
+
+        protected override bool ScrollToPosition(ICaretPosition caret)
+        {
+            // should scroll to the target position on every case.
+            return true;
+        }
+
+        protected override int SkipRows()
+        {
+            // it's a fixed number for now.
+            return 3;
+        }
+
+        protected override DrawableLyricListItem CreateLyricListItem(Lyric item)
+            => new DrawableDetailLyricListItem(item);
+
+        protected override Row GetCreateNewLyricRow()
+            => new CreateNewLyricDetailRow();
+
+        public partial class DrawableDetailLyricListItem : DrawableLyricListItem
+        {
+            public DrawableDetailLyricListItem(Lyric item)
+                : base(item)
+            {
+            }
+
+            protected override Row CreateEditRow(Lyric lyric)
+                => new EditLyricDetailRow(lyric);
+        }
     }
 }
